backend/thorpat/api/v1/activitylog/views.py
from rest_framework import viewsets
from thorpat.apps.activitylog.models import ActivityLog
from .serializers import ActivityLogSerializer

class ActivityLogViewSet(viewsets.ModelViewSet):
    queryset = ActivityLog.objects.all()
    serializer_class = ActivityLogSerializer
