requirements.txt
django
psycopg2-binary
psycopg[pool]
ipython
django-ninja
PyJWT 
cryptography 
django-countries
django-phonenumber-field
phonenumberslite
djangorestframework
sentry-sdk[django]
colorlog
django-typer
django-prometheus
pytest  # If you want to pin a specific version
pytest-django
pytest-xdist
factory-boy
debugpy
pytest-cov
flake8-pyproject
django-cors-headers
django-allauth
requests
drf-yasg
djangorestframework
markdown
django-filter
drf-spectacular
djangorestframework-simplejwt
celery
redis
django-cacheops
dj-rest-auth[with_social]
django-widget-tweaks
django-htmx
pillow
